Implement terminal and browser clients over the shared core
rh-terminal: Bevy 0.18 + bevy_ratatui presentation with keyboard and
mouse (hover inspection, click-to-act), modal overlays, styled event
log, splash/grimoire/relationships/valley/case-report screens, and
share-code save files. rh-web: purpose-built Canvas/HTML WASM client
with the identical intent surface, localStorage replay saves, clipboard
share codes, and a state-digest accessor. Verified in-browser: a full
winning run pasted as a share code replays to Victory with a digest
byte-identical to the native replay.

+++ crates/rh-core/src/viability.rs
@@ -164,7 +164,7 @@ pub fn hunt_viability(
     let effective_hp = u32::from(loadout.hunter_hp + draught_heal * loadout.draughts) * 1000;
     let mut survive = (effective_hp / incoming) as i32;
     survive += i32::from(loadout.physical.min(2)) * 3; // snare denial
-    survive -= i32::from(loadout.draughts) as i32; // drinking costs actions
+    survive -= i32::from(loadout.draughts); // drinking costs actions
     if loadout.on_consecrated_ground {
         survive += 1; // the ward burns the revenant even as it approaches
     }
